+++ packages/utils/interactive-hooks/src/useKeyProps.ts
@@ -1,14 +1,106 @@
-import { KeyCallback, KeyPressProps } from './useKeyProps.types';
+import * as React from 'react';
+import { Platform } from 'react-native';
 import { memoize } from '@fluentui-react-native/memo-cache';
+import { KeyCallback, KeyPressEvent, KeyPressProps } from './useKeyProps.types';
 
-function getKeyUpPropsWorker(_userCallback: KeyCallback, ..._keys: string[]): KeyPressProps {
-  // No keyboard event handling support in React Native on iOS or Android
-  return {};
+/**
+ * Verifies if nativeEvent contains modifier key.
+ * @param nativeEvent
+ * @returns `true` if one or more of modifier keys are `true`
+ */
+export const isModifierKey = (nativeEvent: any): boolean => {
+  return (
+    nativeEvent &&
+    (nativeEvent.alt ||
+      nativeEvent.altKey ||
+      nativeEvent.ctrl ||
+      nativeEvent.ctrlKey ||
+      nativeEvent.meta ||
+      nativeEvent.metaKey ||
+      nativeEvent.shift ||
+      nativeEvent.shiftKey)
+  );
+};
+
+/**
+ * Re-usable hook for an onKeyDown event.
+ * @param userCallback The function you want to be called once the key has been activated on key up
+ * @param keys A string of the key you want to perform some action on. If undefined, always invokes userCallback
+ * @returns onKeyEvent() - Callback to determine if key was pressed, if so, call userCallback
+ * @deprecated use the hook `useKeyProps` instead
+ */
+export function useKeyCallback(userCallback?: KeyCallback, ...keys: string[]) {
+  const onKeyEvent = React.useCallback(
+    (e: KeyPressEvent) => {
+      if (userCallback !== undefined && (keys === undefined || keys.includes(e.nativeEvent.key))) {
+        userCallback(e);
+        e.stopPropagation();
+      }
+    },
+    [keys, userCallback],
+  );
+
+  return onKeyEvent;
 }
 
-function getKeyDownPropsWorker(_userCallback: KeyCallback, ..._keys: string[]): KeyPressProps {
-  // No keyboard event handling support in React Native on iOS or Android
-  return {};
+function getKeyCallbackWorker(userCallback?: KeyCallback, ...keys: string[]) {
+  const onKeyEvent = (e: KeyPressEvent) => {
+    if (userCallback !== undefined && !isModifierKey(e.nativeEvent) && (keys === undefined || keys.includes(e.nativeEvent.key))) {
+      userCallback(e);
+      e.stopPropagation();
+    }
+  };
+  return onKeyEvent;
+}
+
+function getKeyUpPropsWorker(userCallback: KeyCallback, ...keys: string[]): KeyPressProps {
+  const keyboardProps = Platform.select({
+    ios: undefined,
+    android: undefined,
+    macos: {
+      onKeyUp: getKeyCallbackWorker(userCallback, ...keys),
+      validKeysUp: keys,
+    },
+    windows: {
+      onKeyUp: getKeyCallbackWorker(userCallback, ...keys),
+      keyUpEvents: keys.map((keyCode) => {
+        return { key: keyCode };
+      }),
+    },
+    // win32
+    default: {
+      onKeyUp: getKeyCallbackWorker(userCallback, ...keys),
+      keyUpEvents: keys.map((keyCode) => {
+        return { key: keyCode };
+      }),
+    },
+  });
+  return keyboardProps;
+}
+
+function getKeyDownPropsWorker(userCallback: KeyCallback, ...keys: string[]): KeyPressProps {
+  const keyboardProps = Platform.select({
+    ios: undefined,
+    android: undefined,
+    macos: {
+      onKeyDown: getKeyCallbackWorker(userCallback, ...keys),
+      validKeysDown: keys,
+    },
+    windows: {
+      onKeyDown: getKeyCallbackWorker(userCallback, ...keys),
+      keyDownEvents: keys.map((keyCode) => {
+        return { key: keyCode };
+      }),
+    },
+    // win32
+    default: {
+      onKeyDown: getKeyCallbackWorker(userCallback, ...keys),
+      keyDownEvents: keys.map((keyCode) => {
+        return { key: keyCode };
+      }),
+    },
+  });
+  return keyboardProps;
 }
 
 /**
@@ -34,3 +126,9 @@ export const useKeyDownProps = memoize(getKeyDownPropsWorker);
  * @returns KeyPressProps: An object containing the correct platform specific props to handle key press
  */
 export const useKeyProps = memoize(getKeyUpPropsWorker);
+
+/** Exposes the behavior of useKeyProps for the current platform as a boolean */
+export const preferKeyDownForKeyEvents = Platform.select({
+  macos: true,
+  default: false,
+});
